+++ crates/fuzz/src/fuzz_client.rs
@@ -1,12 +1,13 @@
 #![allow(dead_code)]
 
-use crate::error::*;
 use solana_sdk::account::AccountSharedData;
 use solana_sdk::hash::Hash;
 use solana_sdk::instruction::Instruction;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::Keypair;
 use solana_sdk::sysvar::Sysvar;
+use solana_sdk::transaction::TransactionError;
+
 use trident_config::Config;
 use trident_svm::utils::ProgramEntrypoint;
 
@@ -23,7 +24,7 @@ pub trait FuzzClient {
     fn warp_to_slot(&mut self, warp_slot: u64);
 
     /// Forward in time by the desired number of seconds
-    fn forward_in_time(&mut self, seconds: i64) -> Result<(), FuzzClientError>;
+    fn forward_in_time(&mut self, seconds: i64);
 
     /// Create or overwrite a custom account, subverting normal runtime checks.
     fn set_account_custom(&mut self, address: &Pubkey, account: &AccountSharedData);
@@ -41,7 +42,7 @@ pub trait FuzzClient {
     fn process_instructions(
         &mut self,
         _instructions: &[Instruction],
-    ) -> Result<(), FuzzClientError>;
+    ) -> Result<(), TransactionError>;
 
     // Clear Temp account created during fuzzing iteration
     fn clear_accounts(&mut self);